consolidate parser logic
