+++ packages/fetch/package.json
@@ -14,6 +14,7 @@
     "access": "public"
   },
   "devDependencies": {
+    "@entangled/interface": "^2.0.0-alpha.11",
     "@typescript-eslint/parser": "2.18.0",
     "eslint": "6.8.0",
     "typescript": "3.8.3"
